pyproject.toml
[tool.poetry]
name = "check image"
version = "0.1.0"
description = ""
authors = ["abhinav-essbee <abhinavraj.essbee@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
google = "^3.0.0"
google-generativeai = "^0.8.4"
dotenv = "^0.9.9"
pillow = "^11.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
